remove DefaultHeaderKey
